+++ pkg/initialize/middleware.go
@@ -3,6 +3,7 @@ package initialize
 import (
 	_ "github.com/nite-coder/bifrost/pkg/middleware/addprefix"
 	_ "github.com/nite-coder/bifrost/pkg/middleware/cors"
+	_ "github.com/nite-coder/bifrost/pkg/middleware/iprestriction"
 	_ "github.com/nite-coder/bifrost/pkg/middleware/mirror"
 	_ "github.com/nite-coder/bifrost/pkg/middleware/ratelimit"
 	_ "github.com/nite-coder/bifrost/pkg/middleware/replacepath"